+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-### Unreleased
+### 4.18.0
 - Add `gpt-4.5-preview` model.
 - Add model encoding mappings for `gpt-4.5-preview`, `o1` and `o3` models.
 - Add `-InterruptResponse` parameter to `Set-RealtimeSessionConfiguration`